refactor(hopr-strategy): split channel_lifecycle into mod.rs + config.rs; rename tests
- Convert `channel_lifecycle.rs` → `channel_lifecycle/{mod.rs,config.rs}`:
  all config structs extracted to `config.rs`, re-exported from `mod.rs`
- Rename tests to "should" phrasing without `test_` prefix per Rust guidelines
